feat(condo): DOMA-8414 added video to map creation form
