+++ apps/condo/next.config.js
@@ -51,6 +51,9 @@ const condoRBDomain = conf['CONDORB_DOMAIN']
 const sentryConfig = conf['SENTRY_CONFIG'] ? JSON.parse(conf['SENTRY_CONFIG']) : {}
 const apolloBatchingEnabled = !falsey(conf['APOLLO_BATCHING_ENABLED'])
 const tourConfig = JSON.parse(conf['TOUR_CONFIG'] || '{}')
+const externalGuideUrl = conf['EXTERNAL_GUIDE_URL'] || ''
+const residentAppLandingUrl = conf['RESIDENT_APP_LANDING_URL'] || ''
+const createMapVideoUrl = conf['CREATE_MAP_VIDEO_URL'] || ''
 
 let nextConfig = withTM(withLess(withCSS({
     publicRuntimeConfig: {
@@ -83,6 +86,9 @@ let nextConfig = withTM(withLess(withCSS({
         apolloBatchingEnabled,
         currentVersion: getCurrentVersion(),
         tourConfig,
+        externalGuideUrl,
+        residentAppLandingUrl,
+        createMapVideoUrl,
     },
     lessLoaderOptions: {
         javascriptEnabled: true,
